🍓 Avoid code duplication by extraction task related logic to helper classes
