+++ app/src/main/java/com/example/sunrise/MainActivity.java
@@ -1,53 +1,23 @@
 package com.example.sunrise;
 
-import android.content.res.ColorStateList;
 import android.os.Bundle;
-import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.MenuItem;
 import android.view.View;
-import android.widget.Button;
-import android.widget.PopupMenu;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.navigation.NavController;
 import androidx.navigation.fragment.NavHostFragment;
 import androidx.navigation.ui.AppBarConfiguration;
 import androidx.navigation.ui.NavigationUI;
 
-import com.example.sunrise.models.Tag;
-import com.example.sunrise.models.Task;
-import com.example.sunrise.services.TagService;
-import com.example.sunrise.services.TaskService;
+import com.example.sunrise.helpers.TaskCreationHelper;
 import com.google.android.material.appbar.MaterialToolbar;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
-import com.google.android.material.bottomsheet.BottomSheetDialog;
-import com.google.android.material.chip.Chip;
-import com.google.android.material.chip.ChipGroup;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
-import com.google.android.material.textfield.TextInputEditText;
-import com.google.android.material.textfield.TextInputLayout;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.ValueEventListener;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
 
 public class MainActivity extends AppCompatActivity {
-
-    private Chip priorityChip;
-    private BottomSheetDialog bottomSheetDialog;
     private BottomNavigationView bottomNavigationView;
     private FloatingActionButton fab;
-    TextInputEditText editTitle;
-    TextInputLayout titleInputLayout;
-    private ChipGroup tagChips;
-    private List<String> selectedChipIds;
-    TagService tagService;
+    private TaskCreationHelper taskCreationHelper;
 
 
     @Override
@@ -56,9 +26,7 @@ public class MainActivity extends AppCompatActivity {
         setContentView(R.layout.activity_main);
         setupFabButton();
         setupNavigation();
-
-        selectedChipIds = new ArrayList<String>();
-        tagService = new TagService();
+        taskCreationHelper = new TaskCreationHelper(this);
     }
 
     private void setupNavigation() {
@@ -101,122 +69,6 @@ public class MainActivity extends AppCompatActivity {
 
     private void setupFabButton() {
         fab = findViewById(R.id.fab);
-        fab.setOnClickListener(this::showTaskCreationDialog);
+        fab.setOnClickListener(v -> taskCreationHelper.showTaskCreationDialog(v));
     }
-
-    private void showTaskCreationDialog(View v) {
-        bottomSheetDialog = new BottomSheetDialog(MainActivity.this);
-            View bottomSheetContentView = LayoutInflater.from(MainActivity.this).inflate(R.layout.create_task_bottom_sheet, null);
-        bottomSheetDialog.setContentView(bottomSheetContentView);
-        bottomSheetDialog.show();
-
-        titleInputLayout = bottomSheetContentView.findViewById(R.id.titleInputLayout);
-        editTitle = bottomSheetContentView.findViewById(R.id.title);
-        Button createBtn = bottomSheetContentView.findViewById(R.id.create_btn);
-
-        priorityChip = bottomSheetContentView.findViewById(R.id.priority);
-        priorityChip.setOnClickListener(this::showPriorityPopupMenu);
-
-        // Tags row
-        tagChips = bottomSheetContentView.findViewById(R.id.tagChips);
-        populateTagChips();
-
-        createBtn.setOnClickListener(this::createTask);
-    }
-
-    private void populateTagChips() {
-        tagService.getTags(new ValueEventListener() {
-            @Override
-            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
-                List<Tag> tagList = new ArrayList<>();
-                for (DataSnapshot snapshot : dataSnapshot.getChildren()) {
-                    Tag tag = snapshot.getValue(Tag.class);
-                    tagList.add(tag);
-                }
-
-                for (Tag tag : tagList) {
-                    Chip tagChip = (Chip) LayoutInflater.from(MainActivity.this).inflate(R.layout.filter_tag_chip_layout, null);
-                    tagChip.setText(tag.getTitle());
-                    tagChip.setChipBackgroundColor(ColorStateList.valueOf(tag.getColor()));
-
-                    tagChip.setOnCheckedChangeListener((compoundButton, isChecked) -> {
-                        if (isChecked) {
-                            selectedChipIds.add(tag.getTagId());
-                        } else {
-                            selectedChipIds.remove(tag.getTagId());
-                        }
-                    });
-
-                    tagChips.addView(tagChip);
-                }
-            }
-
-            @Override
-            public void onCancelled(@NonNull DatabaseError error) {
-                Log.e("MainActivity", "fetch tags failed");
-            }
-        });
-    }
-
-    private void showPriorityPopupMenu(View view) {
-        PopupMenu popup = new PopupMenu(this, view);
-        popup.getMenuInflater().inflate(R.menu.priorities_menu, popup.getMenu());
-        popup.setOnMenuItemClickListener(this::onPriorityMenuItemClick);
-        popup.show();
-    }
-
-    private void createTask(View view) {
-        String title = Objects.requireNonNull(editTitle.getText()).toString();
-        String priority = getPriorityValue(priorityChip.getText().toString());
-        String userId = Objects.requireNonNull(FirebaseAuth.getInstance().getCurrentUser()).getUid();
-
-        if (title.isEmpty()) {
-            titleInputLayout.setError("Please type title");
-            return;
-        }
-
-        Task task = new Task(title, priority, selectedChipIds, userId);
-
-        // Initialize TaskService to interact with Firebase database
-        TaskService taskService = new TaskService();
-
-        // Save the newly created task to Firebase database
-        taskService.saveTask(task);
-
-        // Dismiss the bottom sheet dialog after task creation
-        bottomSheetDialog.dismiss();
-    }
-
-    private boolean onPriorityMenuItemClick(@NonNull MenuItem item) {
-        int itemId = item.getItemId();
-
-        if (itemId == R.id.priority_high) {
-            priorityChip.setText(R.string.priority_high);
-            return true;
-        } else if (itemId == R.id.priority_medium) {
-            priorityChip.setText(R.string.priority_medium);
-            return true;
-        } else if (itemId == R.id.priority_low) {
-            priorityChip.setText(R.string.priority_low);
-            return true;
-        } else if (itemId == R.id.priority_regular) {
-            priorityChip.setText(R.string.priority_regular);
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    private String getPriorityValue(String selectedPriority) {
-        // Check if the selected priority is the default one
-        if (selectedPriority.equals(getString(R.string.priority))) {
-            // If so, set the priority to the regular
-            return getString(R.string.priority_regular);
-        } else {
-            // Otherwise, return the selected priority
-            return selectedPriority;
-        }
-    }
-
-
 }
